pyproject.toml
[tool.poetry]
name = "slide-gen"
version = "0.1.0"
description = ""
authors = ["Alex Jercan <jercan_alex27@yahoo.com>"]
readme = "README.md"
packages = [{include = "slide_gen"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
fakeyou = "^1.2.5"
ffmpeg-python = "^0.2.0"
flask = "^2.3.2"
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.7.0"

[tool.poetry.scripts]
slide-gen = "slide_gen.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
